Refactored mediaplayer structure
1. removed tash postfix used in mediaplayer.
2. separated the mediaplyer usgae example and moved it to the framework directory.
3. removed .gitignore file.
