+++ apps/examples/mediaplayer/Kconfig
@@ -3,18 +3,18 @@
 # see kconfig-language at https://www.kernel.org/doc/Documentation/kbuild/kconfig-language.txt
 #
 
-config EXAMPLES_MEDIAPLAYER_TASH
+config EXAMPLES_MEDIAPLAYER
 	bool "\"MediaPlayer\" example in tash"
 	default n
 	depends on TASH
 	---help---
 		Enable the \"MediaPlayer\" example in tash
 
-if EXAMPLES_MEDIAPLAYER_TASH
+if EXAMPLES_MEDIAPLAYER
 
-config EXAMPLES_MEDIAPLAYER_TASH_PROGNAME
+config EXAMPLES_MEDIAPLAYER_PROGNAME
 	string "Program name"
-	default "mediaplayer_tash"
+	default "mediaplayer"
 	depends on BUILD_KERNEL
 	---help---
 		This is the name of the program that will be use when the NSH ELF
@@ -24,4 +24,4 @@ endif
 
 config USER_ENTRYPOINT
 	string
-	default "mediaplayer_tash_main" if ENTRY_MEDIAPLAYER_TASH
+	default "mediaplayer_main" if ENTRY_MEDIAPLAYER
