+++ apps/examples/mediaplayer/Kconfig_ENTRY
@@ -1,0 +1,3 @@
+config ENTRY_MEDIAPLAYER
+	bool "\"MediaPlayer\" example in tash"
+	depends on EXAMPLES_MEDIAPLAYER
