--- apps/examples/mediaplayer_tash/.gitignore
@@ -1,11 +1,0 @@
-/Make.dep
-/.depend
-/.built
-/*.asm
-/*.obj
-/*.rel
-/*.lst
-/*.sym
-/*.adb
-/*.lib
-/*.src
