--- apps/examples/mediaplayer_tash/Kconfig_ENTRY
@@ -1,3 +1,0 @@
-config ENTRY_MEDIAPLAYER_TASH
-	bool "\"MediaPlayer\" example in tash"
-	depends on EXAMPLES_MEDIAPLAYER_TASH
